Move HackneyLogo into Header component folder
